Avoid using deflated streams in Unity Core, and move pixel helper methods to ImageBase. Issue #257
