+++ DICOM/IO/Writer/DicomFileWriter.cs
@@ -3,6 +3,7 @@
 
 namespace Dicom.IO.Writer
 {
+    using System;
     using System.IO;
     using System.IO.Compression;
 
@@ -154,6 +155,9 @@ namespace Dicom.IO.Writer
 
             if (syntax.IsDeflate)
             {
+#if NET35
+                throw new NotSupportedException("Deflated datasets not supported in Unity.");
+#else
                 using (var uncompressed = new MemoryStream())
                 {
                     var temp = new StreamByteTarget(uncompressed);
@@ -167,6 +171,7 @@ namespace Dicom.IO.Writer
                         target.Write(compressed.ToArray(), 0, (uint)compressed.Length);
                     }
                 }
+#endif
             }
             else
             {
@@ -264,6 +269,6 @@ namespace Dicom.IO.Writer
             }
         }
 
-        #endregion
+#endregion
     }
 }
