+++ DICOM/Imaging/WindowsImage.cs
@@ -21,7 +21,6 @@ namespace Dicom.Imaging
     {
         /// <summary>
         /// Convenience method to access UWP <see cref="IImage"/> instance as UWP <see cref="WriteableBitmap"/>.
-        /// The returned <see cref="WriteableBitmap"/> will be disposed when the <see cref="IImage"/> is disposed.
         /// </summary>
         /// <param name="image"><see cref="IImage"/> object.</param>
         /// <returns><see cref="WriteableBitmap"/> contents of <paramref name="image"/>.</returns>
@@ -118,117 +117,6 @@ namespace Dicom.Imaging
                 this.height,
                 new PinnedIntArray(this.pixels.Data),
                 this.image == null ? null : Create(this.image));
-        }
-
-        private static byte[] ToBytes(
-            ref int width,
-            ref int height,
-            int components,
-            bool flipX,
-            bool flipY,
-            int rotation,
-            int[] data)
-        {
-            var processed = Rotate(ref width, ref height, rotation, data);
-            processed = Flip(width, height, flipX, flipY, processed);
-
-            // TODO Consider to make use of "components"
-            var length = 4 * width * height;
-            var bytes = new byte[length];
-
-            Buffer.BlockCopy(processed, 0, bytes, 0, length);
-            return bytes;
-        }
-
-        private static int[] Rotate(ref int width, ref int height, int angle, int[] data)
-        {
-            int[] result;
-            angle %= 360;
-
-            var i = 0;
-            if (angle > 0 && angle <= 90)
-            {
-                result = new int[width * height];
-                for (var x = 0; x < width; x++)
-                {
-                    for (var y = height - 1; y >= 0; y--, i++)
-                    {
-                        result[i] = data[y * width + x];
-                    }
-                }
-                var tmp = width;
-                width = height;
-                height = tmp;
-            }
-            else if (angle > 90 && angle <= 180)
-            {
-                result = new int[width * height];
-                for (var y = height - 1; y >= 0; y--)
-                {
-                    for (var x = width - 1; x >= 0; x--, i++)
-                    {
-                        result[i] = data[y * width + x];
-                    }
-                }
-            }
-            else if (angle > 180 && angle <= 270)
-            {
-                result = new int[width * height];
-                for (var x = width - 1; x >= 0; x--)
-                {
-                    for (var y = 0; y < height; y++, i++)
-                    {
-                        result[i] = data[y * width + x];
-                    }
-                }
-                var tmp = width;
-                width = height;
-                height = tmp;
-            }
-            else
-            {
-                result = data;
-            }
-            return result;
-        }
-
-        private static int[] Flip(int w, int h, bool flipX, bool flipY, int[] p)
-        {
-            var i = 0;
-            int[] tmp, result;
-
-            if (flipX)
-            {
-                tmp = new int[w * h];
-                for (var y = h - 1; y >= 0; y--)
-                {
-                    for (var x = 0; x < w; x++, i++)
-                    {
-                        tmp[i] = p[y * w + x];
-                    }
-                }
-            }
-            else
-            {
-                tmp = p;
-            }
-            if (flipY)
-            {
-                result = new int[w * h];
-                for (var y = 0; y < h; y++)
-                {
-                    for (var x = w - 1; x >= 0; x--, i++)
-                    {
-                        result[i] = tmp[y * w + x];
-                    }
-                }
-            }
-            else
-            {
-                result = tmp;
-            }
-
-            return result;
         }
 
         private static WriteableBitmap Create(int width, int height, byte[] bytes)
